// src/common/types.ts
export type UPermission = ValidPermissionKey<PermissionKey>
// ...
export interface IReqApp<TU extends ReqUser = ReqUser> extends Request {
	id: string
	user?: TU
	apiKey?: {
		id: string
		name: string
		userId: string
	}
}
// ...
export interface IPagingData<T> {
	docs: T[]
	count: number
}

export interface ReqUser
	extends Pick<
		User,
		| 'id'
		| 'username'
		| 'mfaTelegramEnabled'
		| 'mfaTotpEnabled'
		| 'totpSecret'
		| 'telegramUsername'
		| 'enabled'
		| 'created'
		| 'modified'
	> {
	sessionId: string
	password: string
	permissions: UPermission[]
}